+++ CHANGELOG.md
@@ -22,4 +22,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated 2 specifications (main)
 - Updated task plan and task index (main)
 - Updated specification `quality-pipeline` (main)
-
+- Completed task `phase-1` (main)